pyproject.toml
[project]
name = "sosediapi"
version = "0.1.0"
description = "Sosedi API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]==5.3.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
